+++ apps/laboratory/tests/shared/utils/project.ts
@@ -2,11 +2,12 @@ import { devices } from '@playwright/test'
 
 import { DESKTOP_DEVICES, MOBILE_DEVICES } from '@reown/appkit-testing'
 
-const LIBRARIES = ['ethers', 'ethers5', 'wagmi', 'solana', 'bitcoin'] as const
+const LIBRARIES = ['ethers', 'wagmi'] as const
 const MULTICHAIN_LIBRARIES = [
   'multichain-no-adapters',
   'multichain-all',
   'multichain-all-email',
+  'multi-wallet-multichain',
   'multichain-ethers-solana',
   'multichain-ethers5-solana',
   'multichain-wagmi-solana'
@@ -77,7 +78,10 @@ const SINGLE_ADAPTER_EVM_TESTS = [
   'wallet.spec.ts',
   'wallet-button.spec',
   'verify.spec.ts',
-  'email-after-farcaster.spec.ts'
+  'email-after-farcaster.spec.ts',
+  'multi-wallet.spec.ts',
+  'multi-wallet-siwx.spec.ts',
+  'flag-enable-reconnect.spec.ts'
 ]
 
 const CORE_TESTS = ['sign-client.spec.ts', 'universal-provider.spec.ts', 'core.spec.ts']
@@ -92,7 +96,8 @@ const SINGLE_ADAPTER_SOLANA_TESTS = [
   'no-email.spec.ts',
   'no-socials.spec.ts',
   'wallet.spec.ts',
-  'wallet-button.spec'
+  'wallet-button.spec',
+  'multi-wallet.spec.ts'
 ]
 
 const CLOUD_AUTH_TESTS = ['cloud-auth.spec.ts']
@@ -158,6 +163,12 @@ const customProjectProperties: CustomProjectProperties = {
   },
   'Desktop Chrome/multichain-all': {
     testMatch: /^.*\/multichain-all\.spec\.ts$/u
+  },
+  'Desktop Firefox/multi-wallet-multichain': {
+    testMatch: /^.*\/multi-wallet-multichain\.spec\.ts$/u
+  },
+  'Desktop Chrome/multi-wallet-multichain': {
+    testMatch: /^.*\/multi-wallet-multichain\.spec\.ts$/u
   },
   'Desktop Firefox/multichain-ethers-solana': {
     testMatch: /^.*\/multichain-ethers-.*\.spec\.ts$/u