+++ apps/laboratory/tests/siwe-email.spec.ts
@@ -113,8 +113,8 @@ emailSiweTest(
   }
 )
 
-emailSiweTest('it should upgrade wallet', async ({ library }) => {
-  const walletUpgradePage = await page.clickWalletUpgradeCard(context, library)
+emailSiweTest('it should upgrade wallet', async () => {
+  const walletUpgradePage = await page.clickWalletUpgradeCard(context)
   expect(walletUpgradePage.url()).toContain(SECURE_WEBSITE_URL)
   await walletUpgradePage.close()
   await page.closeModal()
@@ -163,7 +163,8 @@ emailSiweTest('it should fallback to the last session when cancel siwe from AppK
 })
 
 emailSiweTest('it should disconnect correctly', async () => {
-  await page.goToSettings()
+  await page.goToProfileWalletsView()
+  await page.clickProfileWalletsMoreButton()
   await page.disconnect()
   await validator.expectDisconnected()
 })