+++ apps/laboratory/tests/smart-account.spec.ts
@@ -8,6 +8,7 @@ import { EOA, ModalWalletValidator, SMART_ACCOUNT } from './shared/validators/Mo
 let page: ModalWalletPage
 let validator: ModalWalletValidator
 let context: BrowserContext
+let tempEmail: string
 /* eslint-enable init-declarations */
 
 // -- Setup --------------------------------------------------------------------
@@ -41,7 +42,7 @@ smartAccountTest.beforeAll(async ({ browser, library }) => {
   await validator.expectSwitchedNetworkOnNetworksView('Polygon')
   await page.closeModal()
 
-  const tempEmail = await email.getEmailAddressToUse()
+  tempEmail = await email.getEmailAddressToUse()
   await page.emailFlow({ emailAddress: tempEmail, context, mailsacApiKey })
 
   await validator.expectConnected()
@@ -54,10 +55,8 @@ smartAccountTest.afterAll(async () => {
 // -- Tests --------------------------------------------------------------------
 smartAccountTest('it should use a smart account', async () => {
   await validator.expectConnected()
-  await page.openAccount()
-  await validator.expectActivateSmartAccountPromoVisible(false)
-  await page.openProfileView()
-  await page.openSettings()
+  await page.openProfileWalletsView()
+  await page.clickProfileWalletsMoreButton()
   await validator.expectChangePreferredAccountToShow(EOA)
   await page.closeModal()
 })
@@ -87,8 +86,8 @@ smartAccountTest(
     await page.closeModal()
     await validator.expectAccountButtonReady()
 
-    await page.openAccount()
-    await page.openProfileView()
+    await page.openProfileWalletsView()
+    await page.clickProfileWalletsMoreButton()
     await validator.expectTogglePreferredTypeVisible(false)
     await page.closeModal()
     await validator.expectAccountButtonReady()
@@ -108,7 +107,8 @@ smartAccountTest('it should switch to smart account and sign', async ({ library 
   await page.closeModal()
   await validator.expectAccountButtonReady()
 
-  await page.goToSettings()
+  await page.openProfileWalletsView()
+  await page.clickProfileWalletsMoreButton()
   await validator.expectChangePreferredAccountToShow(SMART_ACCOUNT)
   await page.togglePreferredAccountType()
   await validator.expectChangePreferredAccountToShow(EOA)
@@ -129,7 +129,8 @@ smartAccountTest('it should switch to smart account and sign', async ({ library 
 smartAccountTest('it should switch to eoa and sign', async ({ library }) => {
   const namespace = library === 'solana' ? 'solana' : 'eip155'
 
-  await page.goToSettings()
+  await page.openProfileWalletsView()
+  await page.clickProfileWalletsMoreButton()
   await validator.expectChangePreferredAccountToShow(EOA)
   await page.togglePreferredAccountType()
   await validator.expectChangePreferredAccountToShow(SMART_ACCOUNT)
@@ -150,14 +151,16 @@ smartAccountTest(
       return
     }
 
-    await page.goToSettings()
+    await page.openProfileWalletsView()
+    await page.clickProfileWalletsMoreButton()
     await validator.expectChangePreferredAccountToShow(SMART_ACCOUNT)
     await page.closeModal()
 
     await page.page.reload()
     await validator.expectAccountButtonReady()
 
-    await page.goToSettings()
+    await page.openProfileWalletsView()
+    await page.clickProfileWalletsMoreButton()
     await validator.expectChangePreferredAccountToShow(SMART_ACCOUNT)
     await page.closeModal()
   }
@@ -172,21 +175,47 @@ smartAccountTest(
       return
     }
 
-    await page.goToSettings()
+    await page.openProfileWalletsView()
+    await page.clickProfileWalletsMoreButton()
     await page.togglePreferredAccountType()
     await validator.expectChangePreferredAccountToShow(EOA)
 
     await page.page.reload()
     await validator.expectAccountButtonReady()
 
-    await page.goToSettings()
+    await page.openProfileWalletsView()
+    await page.clickProfileWalletsMoreButton()
     await validator.expectChangePreferredAccountToShow(EOA)
     await page.closeModal()
   }
 )
 
 smartAccountTest('it should disconnect correctly', async () => {
-  await page.goToSettings()
+  await page.openProfileWalletsView()
+  await page.clickProfileWalletsMoreButton()
+  await page.disconnect()
+  await validator.expectDisconnected()
+})
+
+smartAccountTest('it should be able to switch after disconnecting', async ({ library }) => {
+  const mailsacApiKey = process.env['MAILSAC_API_KEY']
+  if (!mailsacApiKey) {
+    throw new Error('MAILSAC_API_KEY is not set')
+  }
+
+  const namespace = library === 'solana' ? 'solana' : 'eip155'
+
+  if (namespace !== 'eip155') {
+    return
+  }
+
+  await page.emailFlow({ emailAddress: tempEmail, context, mailsacApiKey })
+  await validator.expectConnected()
+
+  await page.openProfileWalletsView()
+  await page.clickProfileWalletsMoreButton()
+  await page.togglePreferredAccountType()
+  await validator.expectChangePreferredAccountToShow(SMART_ACCOUNT)
   await page.disconnect()
   await validator.expectDisconnected()
 })