+++ packages/adapters/wagmi/src/connectors/AuthConnector.ts
@@ -3,16 +3,17 @@ import { SwitchChainError, getAddress } from 'viem'
 import type { Address } from 'viem'
 
 import {
+  type ChainNamespace,
   ConstantsUtil as CommonConstantsUtil,
   ConstantsUtil,
   type EmbeddedWalletTimeoutReason
 } from '@reown/appkit-common'
 import { NetworkUtil } from '@reown/appkit-common'
 import {
-  AccountController,
   AlertController,
   ChainController,
-  ConnectorController
+  ConnectorController,
+  getPreferredAccountType
 } from '@reown/appkit-controllers'
 import { ErrorUtil } from '@reown/appkit-utils'
 import { W3mFrameProvider } from '@reown/appkit-wallet'
@@ -67,7 +68,9 @@ export function authConnector(parameters: AuthParameters) {
             AlertController.open(ErrorUtil.ALERT_ERRORS.UNVERIFIED_DOMAIN, 'error')
           }
         },
-        abortController: ErrorUtil.EmbeddedWalletAbortController
+        abortController: ErrorUtil.EmbeddedWalletAbortController,
+        getActiveCaipNetwork: (namespace?: ChainNamespace) =>
+          ChainController.getActiveCaipNetwork(namespace)
       })
     }
 
@@ -95,7 +98,7 @@ export function authConnector(parameters: AuthParameters) {
       }
     }
 
-    const preferredAccountType = AccountController.state.preferredAccountTypes?.eip155
+    const preferredAccountType = getPreferredAccountType('eip155')
 
     const {
       address,
@@ -175,7 +178,9 @@ export function authConnector(parameters: AuthParameters) {
             } else if (reason === 'unverified_domain') {
               AlertController.open(ErrorUtil.ALERT_ERRORS.UNVERIFIED_DOMAIN, 'error')
             }
-          }
+          },
+          getActiveCaipNetwork: (namespace?: ChainNamespace) =>
+            ChainController.getActiveCaipNetwork(namespace)
         })
       }
 
@@ -213,7 +218,7 @@ export function authConnector(parameters: AuthParameters) {
         }
         const provider = await this.getProvider()
 
-        const preferredAccountType = AccountController.state.preferredAccountTypes?.eip155
+        const preferredAccountType = getPreferredAccountType('eip155')
 
         // We connect instead, since changing the chain may cause the address to change as well
         const response = await provider.connect({