+++ packages/appkit/src/client/appkit.ts
@@ -18,7 +18,8 @@ import {
   type Features,
   type Metadata,
   PublicStateController,
-  type RemoteFeatures
+  type RemoteFeatures,
+  getPreferredAccountType
 } from '@reown/appkit-controllers'
 import {
   AccountController,
@@ -62,6 +63,43 @@ export class AppKit extends AppKitBaseClient {
   private authProvider?: W3mFrameProvider
 
   // -- Private ------------------------------------------------------------------
+
+  private onAuthProviderConnected(user: W3mFrameTypes.Responses['FrameGetUserResponse']) {
+    const namespace = ChainController.state.activeChain as ChainNamespace
+
+    // To keep backwards compatibility, eip155 chainIds are numbers and not actual caipChainIds
+    const caipAddress =
+      namespace === ConstantsUtil.CHAIN.EVM
+        ? (`eip155:${user.chainId}:${user.address}` as CaipAddress)
+        : (`${user.chainId}:${user.address}` as CaipAddress)
+
+    const defaultAccountType = OptionsController.state.defaultAccountTypes[namespace]
+    const currentAccountType = getPreferredAccountType(namespace)
+    const preferredAccountType =
+      (user.preferredAccountType as W3mFrameTypes.AccountType) ||
+      currentAccountType ||
+      defaultAccountType
+
+    /*
+     * This covers the case where user switches back from a smart account supported
+     *  network to a non-smart account supported network resulting in a different address
+     */
+
+    if (!HelpersUtil.isLowerCaseMatch(user.address, AccountController.state.address)) {
+      this.syncIdentity({
+        address: user.address,
+        chainId: user.chainId,
+        chainNamespace: namespace
+      })
+    }
+
+    this.setCaipAddress(caipAddress, namespace)
+    this.setUser({ ...(AccountController.state.user || {}), ...user }, namespace)
+    this.setSmartAccountDeployed(Boolean(user.smartAccountDeployed), namespace)
+    this.setPreferredAccountType(preferredAccountType, namespace)
+
+    this.setLoading(false, namespace)
+  }
   private setupAuthConnectorListeners(provider: W3mFrameProvider) {
     provider.onRpcRequest((request: W3mFrameTypes.RPCRequest) => {
       if (W3mFrameHelpers.checkIfRequestExists(request)) {
@@ -113,78 +151,14 @@ export class AppKit extends AppKitBaseClient {
       const namespace = ChainController.state.activeChain as ChainNamespace
       const connectorId = ConnectorController.getConnectorId(namespace)
       const isConnectedWithAuth = connectorId === ConstantsUtil.CONNECTOR_ID.AUTH
+
       if (isConnectedWithAuth) {
         this.setCaipAddress(undefined, namespace)
         this.setLoading(false, namespace)
       }
     })
-    provider.onConnect(user => {
-      const namespace =
-        typeof user?.chainId === 'number' ? ConstantsUtil.CHAIN.EVM : ConstantsUtil.CHAIN.SOLANA
-
-      // To keep backwards compatibility, eip155 chainIds are numbers and not actual caipChainIds
-      const caipAddress =
-        namespace === ConstantsUtil.CHAIN.EVM
-          ? (`eip155:${user.chainId}:${user.address}` as CaipAddress)
-          : (`${user.chainId}:${user.address}` as CaipAddress)
-
-      const defaultAccountType = OptionsController.state.defaultAccountTypes[namespace]
-      const currentAccountType = AccountController.state.preferredAccountTypes?.[namespace]
-      const preferredAccountType =
-        (user.preferredAccountType as W3mFrameTypes.AccountType) ||
-        currentAccountType ||
-        defaultAccountType
-
-      /*
-       * This covers the case where user switches back from a smart account supported
-       *  network to a non-smart account supported network resulting in a different address
-       */
-
-      if (!HelpersUtil.isLowerCaseMatch(user.address, AccountController.state.address)) {
-        this.syncIdentity({
-          address: user.address,
-          chainId: user.chainId,
-          chainNamespace: namespace
-        })
-      }
-      this.setCaipAddress(caipAddress, namespace)
-      this.setUser({ ...(AccountController.state.user || {}), ...user }, namespace)
-      this.setSmartAccountDeployed(Boolean(user.smartAccountDeployed), namespace)
-      this.setPreferredAccountType(preferredAccountType, namespace)
-
-      const userAccounts = user.accounts?.map(account =>
-        CoreHelperUtil.createAccount(
-          namespace,
-          account.address,
-          account.type || currentAccountType || defaultAccountType
-        )
-      )
-
-      this.setAllAccounts(
-        userAccounts || [
-          CoreHelperUtil.createAccount(
-            namespace,
-            user.address,
-            (user.preferredAccountType as W3mFrameTypes.AccountType) || preferredAccountType
-          )
-        ],
-        namespace
-      )
-
-      this.setLoading(false, namespace)
-    })
-    provider.onSocialConnected(({ userName }) => {
-      this.setUser(
-        { ...(AccountController.state.user || {}), username: userName },
-        ChainController.state.activeChain
-      )
-    })
-    provider.onGetSmartAccountEnabledNetworks(networks => {
-      this.setSmartAccountEnabledNetworks(
-        networks,
-        ChainController.state.activeChain as ChainNamespace
-      )
-    })
+    provider.onConnect(this.onAuthProviderConnected.bind(this))
+    provider.onSocialConnected(this.onAuthProviderConnected.bind(this))
     provider.onSetPreferredAccount(({ address, type }) => {
       if (!address) {
         return
@@ -195,6 +169,29 @@ export class AppKit extends AppKitBaseClient {
         ChainController.state.activeChain as ChainNamespace
       )
     })
+  }
+
+  private async syncAuthConnectorTheme(provider: W3mFrameProvider | undefined) {
+    if (!provider) {
+      return
+    }
+
+    const theme = ThemeController.getSnapshot()
+    const options = OptionsController.getSnapshot()
+
+    await Promise.all([
+      provider.syncDappData({
+        metadata: options.metadata as Metadata,
+        sdkVersion: options.sdkVersion,
+        projectId: options.projectId,
+        sdkType: options.sdkType
+      }),
+      provider.syncTheme({
+        themeMode: theme.themeMode,
+        themeVariables: theme.themeVariables,
+        w3mThemeVariables: getW3mThemeVariables(theme.themeVariables, theme.themeMode)
+      })
+    ])
   }
 
   private async syncAuthConnector(provider: W3mFrameProvider, chainNamespace: ChainNamespace) {
@@ -221,26 +218,14 @@ export class AppKit extends AppKitBaseClient {
 
     const { isConnected } = await provider.isConnected()
 
-    const theme = ThemeController.getSnapshot()
-    const options = OptionsController.getSnapshot()
-
-    await Promise.all([
-      provider.syncDappData({
-        metadata: options.metadata as Metadata,
-        sdkVersion: options.sdkVersion,
-        projectId: options.projectId,
-        sdkType: options.sdkType
-      }),
-      provider.syncTheme({
-        themeMode: theme.themeMode,
-        themeVariables: theme.themeVariables,
-        w3mThemeVariables: getW3mThemeVariables(theme.themeVariables, theme.themeMode)
-      })
-    ])
-
-    await provider.getSmartAccountEnabledNetworks()
+    await this.syncAuthConnectorTheme(provider)
 
     if (chainNamespace && isAuthSupported && shouldSync) {
+      const enabledNetworks = await provider.getSmartAccountEnabledNetworks()
+      ChainController.setSmartAccountEnabledNetworks(
+        enabledNetworks?.smartAccountEnabledNetworks || [],
+        chainNamespace
+      )
       if (isConnected && this.connectionControllerClient?.connectExternal) {
         await this.connectionControllerClient?.connectExternal({
           id: ConstantsUtil.CONNECTOR_ID.AUTH,
@@ -364,7 +349,9 @@ export class AppKit extends AppKitBaseClient {
           } else if (reason === 'unverified_domain') {
             AlertController.open(ErrorUtil.ALERT_ERRORS.UNVERIFIED_DOMAIN, 'error')
           }
-        }
+        },
+        getActiveCaipNetwork: (namespace?: ChainNamespace) =>
+          ChainController.getActiveCaipNetwork(namespace)
       })
       PublicStateController.subscribeOpen(isOpen => {
         if (!isOpen && this.isTransactionStackEmpty()) {
@@ -372,10 +359,13 @@ export class AppKit extends AppKitBaseClient {
         }
       })
     }
+    const shouldSyncAccount =
+      chainNamespace === ChainController.state.activeChain &&
+      OptionsController.state.enableReconnect
 
-    const shouldSync = chainNamespace === ChainController.state.activeChain
-
-    if (this.authProvider && shouldSync) {
+    if (OptionsController.state.enableReconnect === false) {
+      this.syncAuthConnectorTheme(this.authProvider)
+    } else if (this.authProvider && shouldSyncAccount) {
       this.syncAuthConnector(this.authProvider, chainNamespace)
       this.checkExistingTelegramSocialConnection(chainNamespace)
     }
@@ -474,6 +464,17 @@ export class AppKit extends AppKitBaseClient {
           })
         }
       } else if (newNamespaceProviderType === UtilConstantsUtil.CONNECTOR_TYPE_WALLET_CONNECT) {
+        /*
+         * Switch network for adapters, as each manages
+         * its own connections state and requires network updates
+         */
+        if (!ChainController.state.noAdapters) {
+          const adapter = this.getAdapter(networkNamespace)
+          const provider = ProviderUtil.getProvider(networkNamespace)
+          const providerType = ProviderUtil.getProviderId(networkNamespace)
+
+          await adapter?.switchNetwork({ caipNetwork, provider, providerType })
+        }
         this.setCaipNetwork(caipNetwork)
         this.syncWalletConnectAccount()
       } else {