+++ packages/controllers/src/controllers/ChainController.ts
@@ -16,7 +16,8 @@ import { StorageUtil } from '../utils/StorageUtil.js'
 import type {
   AdapterNetworkState,
   ChainAdapter,
-  NetworkControllerClient
+  NetworkControllerClient,
+  PreferredAccountTypes
 } from '../utils/TypeUtil.js'
 import { withErrorBoundary } from '../utils/withErrorBoundary.js'
 import { AccountController, type AccountControllerState } from './AccountController.js'
@@ -35,8 +36,8 @@ const accountState: AccountControllerState = {
   tokenBalance: [],
   smartAccountDeployed: false,
   addressLabels: new Map(),
-  allAccounts: [],
-  user: undefined
+  user: undefined,
+  preferredAccountType: undefined
 }
 
 const networkState: AdapterNetworkState = {
@@ -160,13 +161,20 @@ const controller = {
       const namespaceNetworks = caipNetworks?.filter(
         network => network.chainNamespace === namespace
       )
+
+      const storedAccountTypes = StorageUtil.getPreferredAccountTypes() || {}
+      const defaultTypes = { ...OptionsController.state.defaultAccountTypes, ...storedAccountTypes }
+
       ChainController.state.chains.set(namespace as ChainNamespace, {
         namespace,
         networkState: proxy({
           ...networkState,
           caipNetwork: namespaceNetworks?.[0]
         }),
-        accountState: proxy(accountState),
+        accountState: proxy({
+          ...accountState,
+          preferredAccountType: defaultTypes[namespace] as PreferredAccountTypes[ChainNamespace]
+        }),
         caipNetworks: namespaceNetworks ?? [],
         ...clients
       })
@@ -309,9 +317,6 @@ const controller = {
     replaceState = true
   ) {
     ChainController.setChainAccountData(chain, { [prop]: value }, replaceState)
-    if (prop === 'status' && value === 'disconnected' && chain) {
-      ConnectorController.removeConnectorId(chain)
-    }
   },
 
   setActiveNamespace(chain: ChainNamespace | undefined) {
@@ -555,7 +560,11 @@ const controller = {
     return approvedCaipNetworkIds
   },
 
-  getActiveCaipNetwork() {
+  getActiveCaipNetwork(chainNamespace?: ChainNamespace) {
+    if (chainNamespace) {
+      return state.chains.get(chainNamespace)?.networkState?.caipNetwork
+    }
+
     return state.activeCaipNetwork
   },
 
@@ -638,8 +647,7 @@ const controller = {
   resetNetwork(namespace: ChainNamespace) {
     ChainController.setAdapterNetworkState(namespace, {
       approvedCaipNetworkIds: undefined,
-      supportsAllNetworks: true,
-      smartAccountEnabledNetworks: []
+      supportsAllNetworks: true
     })
   },
 
@@ -649,6 +657,8 @@ const controller = {
     if (!chainToWrite) {
       throw new Error('Chain is required to set account prop')
     }
+
+    const currentAccountType = ChainController.getAccountProp('preferredAccountType', chainToWrite)
 
     state.activeCaipAddress = undefined
     ChainController.setChainAccountData(chainToWrite, {
@@ -663,11 +673,10 @@ const controller = {
       addressExplorerUrl: undefined,
       tokenBalance: [],
       connectedWalletInfo: undefined,
-      preferredAccountTypes: undefined,
+      preferredAccountType: currentAccountType,
       socialProvider: undefined,
       socialWindow: undefined,
       farcasterUrl: undefined,
-      allAccounts: [],
       user: undefined,
       status: 'disconnected'
     })