+++ packages/scaffold-ui/src/partials/w3m-connect-featured-widget/index.ts
@@ -2,8 +2,9 @@ import { LitElement, html } from 'lit'
 import { property } from 'lit/decorators.js'
 import { ifDefined } from 'lit/directives/if-defined.js'
 
+import { ConstantsUtil as CommonConstantsUtil } from '@reown/appkit-common'
 import type { WcWallet } from '@reown/appkit-controllers'
-import { AssetUtil, ConnectorController } from '@reown/appkit-controllers'
+import { AssetUtil, ConnectionController, ConnectorController } from '@reown/appkit-controllers'
 import { customElement } from '@reown/appkit-ui'
 import '@reown/appkit-ui/wui-flex'
 import '@reown/appkit-ui/wui-list-wallet'
@@ -23,6 +24,10 @@ export class W3mConnectFeaturedWidget extends LitElement {
       return null
     }
 
+    const hasWcConnection = ConnectionController.hasAnyConnection(
+      CommonConstantsUtil.CONNECTOR_ID.WALLET_CONNECT
+    )
+
     return html`
       <wui-flex flexDirection="column" gap="xs">
         ${this.wallets.map(
@@ -33,6 +38,7 @@ export class W3mConnectFeaturedWidget extends LitElement {
               name=${wallet.name ?? 'Unknown'}
               @click=${() => this.onConnectWallet(wallet)}
               tabIdx=${ifDefined(this.tabIdx)}
+              ?disabled=${hasWcConnection}
             >
             </wui-list-wallet>
           `